Changed a bunch of classes to use double precision for computations, needed if we're going to scale this up.

+++ Source/Drawable/Plane.cpp
@@ -1,23 +1,23 @@
 #include "Plane.h"
 
-Plane::Plane(const glm::vec3 & normal, const glm::vec3 & point) {
-	x = (double)normal.x;
-	y = (double)normal.y;
-	z = (double)normal.z;
-	d = -(double)glm::dot(normal, point);
+Plane::Plane(const glm::dvec3 & normal, const glm::dvec3 & point) {
+	x = normal.x;
+	y = normal.y;
+	z = normal.z;
+	d = -glm::dot(normal, point);
 }
 
-Plane::Plane(const glm::vec3 & p1, const glm::vec3 & p2, const glm::vec3 & p3) {
-	glm::vec3 v1 = p2 - p1;
-	glm::vec3 v2 = p3 - p1;
-	glm::vec3 normal = glm::normalize(glm::cross(v1, v2));
-	x = (double)normal.x;
-	y = (double)normal.y;
-	z = (double)normal.z;
-	d = -(double)glm::dot(normal, p1);
+Plane::Plane(const glm::dvec3 & p1, const glm::dvec3 & p2, const glm::dvec3 & p3) {
+	glm::dvec3 v1 = p2 - p1;
+	glm::dvec3 v2 = p3 - p1;
+	glm::dvec3 normal = glm::normalize(glm::cross(v1, v2));
+	x = normal.x;
+	y = normal.y;
+	z = normal.z;
+	d = -glm::dot(normal, p1);
 }
 
-double Plane::distance(const glm::vec3 &point) const {
+double Plane::distance(const glm::dvec3 &point) const {
 	return x * point.x + y * point.y + z * point.z + d;
 }
